+++ src/logging.ts
@@ -2,16 +2,16 @@
 Logging facilities for the extension.
 */
 
-type logLevel = "debug" | "warning" | "error";
+type LogLevel = "debug" | "warning" | "error";
 
 
 const logTitle = 'KataPod';
 
 
-export function log(level: logLevel, message: string) {
+export function log(level: LogLevel, message: string) {
 	console.log(buildLogMessage(level, message));
 }
 
-export function buildLogMessage(level: logLevel, message: string) {
+export function buildLogMessage(level: LogLevel, message: string) {
 	return `${logTitle} [${level.toUpperCase()}] ${message}`;
 }
